+++ common.h
@@ -12,7 +12,7 @@
 #define MIN_GTK_RUNTIME(major, minor, micro) (!gtk_check_version(major, minor, micro))
 
 #define APPNAME "pingpath"
-#define VERSION "1.0.12"
+#define VERSION "1.0.13"
 #define APPVER  APPNAME "-" VERSION
 
 extern locale_t locale, localeC;
@@ -467,7 +467,7 @@ gint snprintg(gchar *string, gulong n, const gchar *format, ...);
 GtkListBoxRow* line_row_new(GtkWidget *child, gboolean visible);
 
 void host_free(t_host *host);
-int host_cmp(const void *a, const void *b);
+int host_cmp(const t_host *a, const t_host *b);
 int ref_cmp(const t_ref *a, const t_ref *b);
 t_ref* ref_new(t_hop *hop, int ndx);
 #if defined(DNS_EXTRA_DEBUG) || defined(WHOIS_EXTRA_DEBUG)