+++ stat.c
@@ -24,11 +24,12 @@ int visibles = -1;
 static t_hop hops[MAXTTL];
 
 static void update_addrname(int at, t_host *b) { // addr is mandatory, name not
-  if (!b) return;
+  if (!b)
+    return;
   gboolean done = false;
   int vacant = -1;
   t_hop *hop = &hops[at];
-  for (int i = 0; i < MAXADDR; i++) {
+  for (uint i = 0; i < G_N_ELEMENTS(hop->host); i++) {
     t_host *a = &hop->host[i];
     if ((vacant < 0) && !a->addr)
       vacant = i;
@@ -69,7 +70,8 @@ static void update_addrname(int at, t_host *b) { // addr is mandatory, name not
 }
 
 static void stop_pings_behind(int from, const char *reason) {
-  for (int i = from; i < MAXTTL; i++) pinger_nth_stop(i, reason);
+  for (int i = from; i < MAXTTL/*G_N_ELEMENTS(pings)*/; i++)
+    pinger_nth_stop(i, reason);
 }
 
 static void set_target_at(int at, const char *info) {
@@ -80,29 +82,41 @@ static void set_target_at(int at, const char *info) {
 
 static void uniq_unreach(int at) {
   const char *addr = hops[at].host[0].addr;
-  if (!addr) return;
+  if (!addr)
+    return;
   int last = at;
-  for (; last > 0; last--) {
-    if (hops[last - 1].reach) break;
-    if (STR_NEQ(addr, hops[last - 1].host[0].addr)) break;
-  }
-  if (tgtat != (last + 1)) set_target_at(last + 1, SKIPDUP_HDR);
+  for (; last > 0; last--)
+    if (hops[last - 1].reach || STR_NEQ(addr, hops[last - 1].host[0].addr))
+      break;
+  if (tgtat != (last + 1))
+    set_target_at(last + 1, SKIPDUP_HDR);
 }
 
-#define STAT_AVERAGE(count, avrg, val) { (count)++; \
-  if ((count) > 0) { if ((avrg) < 0) (avrg) = 0; (avrg) += ((val) - (avrg)) / (count); }}
+#define STAT_AVERAGE(count, avrg, val) do { \
+  (count)++;                                \
+  if ((count) > 0) {                        \
+    if ((avrg) < 0) (avrg) = 0;             \
+    (avrg) += ((val) - (avrg)) / (count);   \
+  }                                         \
+} while (0)
 
 enum { PREV, CURR };
 
 static void update_stat(int at, int rtt, t_tseq *mark, uint rxtx) {
-  if (rxtx & RX) hops[at].recv++;
-  if (rxtx & TX) hops[at].sent++;
-  if (rxtx && hops[at].sent) hops[at].loss = (hops[at].sent - hops[at].recv) * 100. / hops[at].sent;
+  if (rxtx & RX)
+    hops[at].recv++;
+  if (rxtx & TX)
+    hops[at].sent++;
+  if (rxtx && hops[at].sent)
+    hops[at].loss = (hops[at].sent - hops[at].recv) * 100. / hops[at].sent;
   if (rtt > 0) {
-    if ((rtt < hops[at].best) || (hops[at].best < 0)) hops[at].best = rtt;
-    if ((rtt > hops[at].wrst) || (hops[at].wrst < 0)) hops[at].wrst = rtt;
+    if ((rtt < hops[at].best) || (hops[at].best < 0))
+      hops[at].best = rtt;
+    if ((rtt > hops[at].wrst) || (hops[at].wrst < 0))
+      hops[at].wrst = rtt;
     STAT_AVERAGE(hops[at].known_rtts, hops[at].avrg, rtt);
-    if (hops[at].prev > 0) STAT_AVERAGE(hops[at].known_jttrs, hops[at].jttr, abs(rtt - hops[at].prev));
+    if (hops[at].prev > 0)
+      STAT_AVERAGE(hops[at].known_jttrs, hops[at].jttr, abs(rtt - hops[at].prev));
     hops[at].prev = hops[at].last;
     hops[at].last = rtt;
   }
@@ -113,16 +127,23 @@ static void update_stat(int at, int rtt, t_tseq *mark, uint rxtx) {
       hops[at].rseq[CURR].rtt = rtt; hops[at].rseq[CURR].seq = mark->seq;
     }
   }
-  if (!pinger_state.gotdata) pinger_state.gotdata = (hops[at].recv > 0);
-  if (rxtx == TX) hops[at].tout = true; else if (rxtx != NONE) hops[at].tout = false;
-  if (!hops[at].tout && (visibles < at) && (at < tgtat)) { visibles = at; pinger_vis_rows(at + 1); }
+  if (!pinger_state.gotdata)
+    pinger_state.gotdata = (hops[at].recv > 0);
+  if (rxtx == TX)
+    hops[at].tout = true;
+  else if (rxtx != NONE)
+    hops[at].tout = false;
+  if (!hops[at].tout && (visibles < at) && (at < tgtat)) {
+    visibles = at;
+    pinger_vis_rows(at + 1);
+  }
 }
 
 static int calc_rtt(int at, t_tseq *mark) {
-  if (!hops[at].mark.sec) return -1;
+  if (!hops[at].mark.sec)
+    return -1;
   int rtt = (mark->sec - hops[at].mark.sec) * G_USEC_PER_SEC + (mark->usec - hops[at].mark.usec);
-  rtt = (rtt > opts.tout_usec) ? -1 : rtt; // skip timeouted responses
-  return rtt;
+  return (rtt > opts.tout_usec) ? -1 : rtt; // skip timeouted responses
 }
 
 // Note: name[0] is shortcut for "" test instead of STR_NEQ(name, UNKN_FIELD)
@@ -133,7 +154,9 @@ static inline const char* addrname(int ndx, t_host *host) {
 }
 
 const char *stat_hop_addrname(int at) {
-  static char hostinfo_cache[MAXTTL][BUFF_SIZE];
+  static char hostinfo_cache[G_N_ELEMENTS(hops)][BUFF_SIZE];
+  if ((at < 0) || ((uint)at >= G_N_ELEMENTS(hops)))
+    return NULL;
   t_hop *hop = &hops[at];
   t_host *host = hop->host;
   if (host[0].addr) { // as a marker
@@ -141,12 +164,14 @@ const char *stat_hop_addrname(int at) {
       return hostinfo_cache[at];
     char *str = hostinfo_cache[at];
     int len = snprintg(hostinfo_cache[at], BUFF_SIZE, "%s", addrname(0, host));
-    if (len > 0) for (int i = 1; (i < MAXADDR) && (len < BUFF_SIZE); i++) {
+    if (len > 0) for (uint i = 1; (i < G_N_ELEMENTS(hop->host)) && (len < BUFF_SIZE); i++) {
       if (host[i].addr) {
         int inc = snprintg(str + len, BUFF_SIZE - len, "\n%s", addrname(i, host));
-        if (inc < 0) break;
+        if (inc < 0)
+          break;
         len += inc;
-      } else break;
+      } else
+        break;
     }
     if (hop->info && (len > 0) && (len < BUFF_SIZE))
       snprintg(str + len, BUFF_SIZE - len, "\n%s", hop->info);
@@ -158,7 +183,9 @@ const char *stat_hop_addrname(int at) {
 }
 
 static const char *host_legend(int at) {
-  static char hostinfo_nl_cache[MAXTTL][MAXHOSTNAME];
+  static char hostinfo_nl_cache[G_N_ELEMENTS(hops)][MAXHOSTNAME];
+  if ((at < 0) || ((uint)at >= G_N_ELEMENTS(hops)))
+    return NULL;
   t_hop *hop = &hops[at];
   t_host *host = hop->host;
   if (host[0].addr) { // as a marker
@@ -181,20 +208,23 @@ static uint column_host(int at, t_ping_column *column, int num G_GNUC_UNUSED) { 
       break;
    column->cell[n] = addrname(n, host);
   }
-  return n;
+  return n; // hosts per hop
 }
 
 static const char *info_whois(int at, int type) {
-  static char whois_cache[MAXTTL][WHOIS_NDX_MAX][BUFF_SIZE];
-  t_whois *whois = hops[at].whois;
+  static char whois_cache[G_N_ELEMENTS(hops)][WHOIS_NDX_MAX][BUFF_SIZE];
+  if ((at < 0) || ((uint)at >= G_N_ELEMENTS(hops)))
+    return NULL;
+  t_hop *hop = &hops[at];
+  t_whois *whois = hop->whois;
   char *view = T_WHOIS_VIEW(whois, type);
   if (!view)
     return NULL;
   char *str = whois_cache[at][type];
   // TODO: reset hops[at].whois_cached[] with changing 'whois_multi'
-  if (!hops[at].whois_cached[type]) {
+  if (!hop->whois_cached[type]) {
     int len = snprintg(str, BUFF_SIZE, "%s", view);
-    if (len > 0) for (uint i = 1; (i < MAXADDR) && (len < BUFF_SIZE); i++) {
+    if (len > 0) for (uint i = 1; (i < G_N_ELEMENTS(hop->whois)) && (len < BUFF_SIZE); i++) {
       view = T_WHOIS_VIEW(&whois[i], type);
       if (!view)
         break;
@@ -203,14 +233,16 @@ static const char *info_whois(int at, int type) {
         break;
       len += inc;
     }
-    hops[at].whois_cached[type] = true;
+    hop->whois_cached[type] = true;
     LOG("%s: [%d,%d] %s", WHOIS_CUP_HDR, at, type, mnemo(str));
   }
   return str;
 }
 
 static const char *whois_legend(int at, int type) {
-  static char whois_nl_cache[MAXTTL][WHOIS_NDX_MAX][MAXHOSTNAME];
+  static char whois_nl_cache[G_N_ELEMENTS(hops)][WHOIS_NDX_MAX][MAXHOSTNAME];
+  if ((at < 0) || ((uint)at >= G_N_ELEMENTS(hops)))
+    return NULL;
   t_whois *whois = hops[at].whois;
   char *view = T_WHOIS_VIEW(whois, type);
   if (view) { // as a marker
@@ -234,10 +266,12 @@ static uint column_whois(int at, t_ping_column *column, int type) { // NONNULL(2
       break;
     column->cell[n] = str;
   }
-  return n;
+  return n; // hosts per hop
 }
 
-static int precision(double val) { return ((val > 0) && (val < 10)) ? ((val < 0.1) ? 2 : 1) : 0; }
+static int precision(double val) {
+  return ((val > 0) && (val < 10)) ? ((val < 0.1) ? 2 : 1) : 0;
+}
 
 static const char* fill_stat_int(int val, char* buff, int size) {
   if (val < 0)
@@ -272,7 +306,8 @@ static void stat_nth_hop_NA(t_hop *hop) { // NA(<0)
 }
 
 static void stat_up_info(int at, const char *info) {
-  if (STR_EQ(hops[at].info, info)) return;
+  if (STR_EQ(hops[at].info, info))
+    return;
   UPD_STR(hops[at].info, info);
   hops[at].dns_cached = false;
   LOG("%s: #%d %s", INFO_HDR, at + 1, info);
@@ -391,10 +426,12 @@ void stat_init(gboolean clean) { // clean start or on reset
     pinger_state.reachable = false;
     memset(hops, 0, sizeof(hops)); // BUFFNOLINT
   }
-  for (int i = 0; i < MAXTTL; i++) {
-    if (clean) hops[i].reach = true;
-    stat_nth_hop_NA(&hops[i]);
-    hops[i].at = i;
+  t_hop *hop = hops;
+  for (uint i = 0; hop && (i < G_N_ELEMENTS(hops)); i++, hop++) {
+    if (clean)
+      hop->reach = true;
+    stat_nth_hop_NA(hop);
+    hop->at = i;
   }
   stat_whois_enabler();
 }
@@ -441,8 +478,10 @@ void stat_save_success(int at, t_ping_success *data) {
 
 void stat_save_discard(int at, t_ping_discard *data) {
   update_addrname(at, &data->host);
-  if ((data->mark.seq - hops[at].mark.seq) != 1) update_stat(at, -1, NULL, RX);
-  else update_stat(at, calc_rtt(at, &data->mark), &data->mark, RX | TX);
+  if ((data->mark.seq - hops[at].mark.seq) != 1)
+    update_stat(at, -1, NULL, RX);
+  else
+    update_stat(at, calc_rtt(at, &data->mark), &data->mark, RX | TX);
   // 'unreach' management
   if (data->reason) {
     int ttl = at + 1;
@@ -473,10 +512,13 @@ void stat_save_info(int at, t_ping_info *data) {
 }
 
 void stat_last_tx(int at) { // update last 'tx' unless done before
-  if (hops[at].tout) { update_stat(at, -1, NULL, TX); hops[at].tout = false; }
+  if (hops[at].tout) {
+    update_stat(at, -1, NULL, TX);
+    hops[at].tout = false;
+  }
 }
 
-double stat_dbl_elem(int at, int type) { // for json (called once)
+double stat_dbl_elem(int at, int type) { // for json resume at exit
   double val = -1;
   switch (type) {
     case PE_LOSS: val = hops[at].loss; break;
@@ -486,7 +528,7 @@ double stat_dbl_elem(int at, int type) { // for json (called once)
   }
   return val;
 }
-int stat_int_elem(int at, int type) { // for json (called once)
+int stat_int_elem(int at, int type) { // for json resume at exit
   int val = -1;
   switch (type) {
     case PE_SENT: val = hops[at].sent; break;
@@ -499,17 +541,20 @@ int stat_int_elem(int at, int type) { // for json (called once)
   return val;
 }
 
-
 #define RSEQ_FOR_SYNC(which) { data->rtt = hops[at].rseq[which].rtt; data->seq = -hops[at].rseq[which].seq; }
 
 void stat_rseq(int at, t_rseq *data) { // assert('at' within range)
   if (data) {
-    if (!data->seq || (data->seq == hops[at].rseq[CURR].seq)) *data = hops[at].rseq[CURR];
-    else if (data->seq == hops[at].rseq[PREV].seq) *data = hops[at].rseq[PREV];
-    else { /* for sync */
-      if (data->seq > hops[at].rseq[CURR].seq) RSEQ_FOR_SYNC(CURR) // rarely, but possible
-      else RSEQ_FOR_SYNC(PREV); // unlikely with iputils-ping
-   }
+    if (!data->seq || (data->seq == hops[at].rseq[CURR].seq))
+      *data = hops[at].rseq[CURR];
+    else if (data->seq == hops[at].rseq[PREV].seq)
+      *data = hops[at].rseq[PREV];
+    else { // for sync
+      if (data->seq > hops[at].rseq[CURR].seq)
+        RSEQ_FOR_SYNC(CURR) // rarely, but possible
+      else
+        RSEQ_FOR_SYNC(PREV); // unlikely with iputils-ping
+    }
   }
 }
 
@@ -537,10 +582,12 @@ void stat_whois_enabler(void) {
 }
 
 void stat_run_whois_resolv(void) {
-  for (int at = opts.range.min; at < opts.range.max; at++) {
-    t_hop *hop = &hops[at];
-    for (int i = 0; i < MAXADDR; i++)
-      if (hop->host[i].addr) whois_resolv(hop, i);
+  t_hop *hop = hops;
+  for (int at = opts.range.min; hop && (at < opts.range.max); at++, hop++) {
+    t_host *host = hop->host;
+    for (uint i = 0; host && (i < G_N_ELEMENTS(hop->host)); i++, host++)
+      if (host->addr)
+        whois_resolv(hop, i);
   }
 }